refactor(landingpage): optimize backer logos
