+++ src/stories/Pages/LandingPage/LandingPage.stories.tsx
@@ -28,10 +28,10 @@ import NFTFiMonoLogo from '../../../assets/LandingPage/supporting/NFTFi-Mono.svg
 import TellerMonoLogo from '../../../assets/LandingPage/supporting/Teller-Mono.svg';
 import NFTAssetPreview from '../../../assets/LandingPage/previews/NFTAsset.svg';
 import NFSCAssetPreview from '../../../assets/LandingPage/previews/NFSCAsset.svg';
-import AlchemyLogo from '../../../assets/LandingPage/backed-by/alchemy.svg';
-import MaskLogo from '../../../assets/LandingPage/backed-by/mask.svg';
-import FoothillLogo from '../../../assets/LandingPage/backed-by/foothill.svg';
-import OrangedaoLogo from '../../../assets/LandingPage/backed-by/orangedao.svg';
+import AlchemyLogo from '../../../assets/LandingPage/backed-by/Alchemy-Poly.png';
+import MaskLogo from '../../../assets/LandingPage/backed-by/Mask-Poly.png';
+import FoothillLogo from '../../../assets/LandingPage/backed-by/Foothill-Poly.png';
+import OrangedaoLogo from '../../../assets/LandingPage/backed-by/OrangeDAO-Poly.png';
 
 const meta = {
 	title: 'Pages/Landingpage',
